Add create fuse button to profile
