+++ src/styles/colors/index.js
@@ -2,4 +2,5 @@ export default {
   redorange: 'rgba(237,92,69,1)',
   grey: 'grey',
   black: 'black',
+  white: 'white',
 };
